[refactor]변수 이름 변경 및 코드 간결화
