--- src/hooks/canvas/useFetchKeyword.js
@@ -1,23 +1,0 @@
-import useFetchDiary from "../diary/queries/useFetchDiary";
-
-const useFetchKeywords = () => {
-  const { diary } = useFetchDiary();
-  const keywords = diary?.keywords ?? [];
-
-  const updatedKeywords =
-    keywords.length === 0
-      ? [
-          {
-            keyword: "자유롭게 그려보세요!",
-            keywordId: 0,
-            imgUrl: null,
-          },
-        ]
-      : [...keywords];
-
-  const isKeywordEmpty = diary?.keywords.length === 0;
-
-  return { keywords: updatedKeywords, isKeywordEmpty };
-};
-
-export default useFetchKeywords;
